# Improve isoformat timestamp performance