@bdraco 2020-06-22
I should add a test to make sure this works with a different timezone as input to the new function in recorder.