@bdraco 2020-06-22
TODO

- [ ] Verified ok with Sqlite
- [ ] Verified ok with MySQL/Maria
- [ ] Verified ok with Postgresql
- [ ] Test for new recorder function added