@bdraco 2020-06-22
returning `None` is not covered.
easy to fix